# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Network Diagnostics Dashboard API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
  version?: string;
  uptime?: number;
  demoMode?: boolean;
}

export interface SystemInfo {
  hostname: string;
  platform: string;
  os: string;
  arch: string;
  uptime: number;
  nodeVersion?: string;
  /** @nullable */
  publicIp?: string | null;
  localIps: string[];
  demoMode: boolean;
  environment: string;
}

export type SystemCapabilitiesCommands = {
  ping: boolean;
  traceroute: boolean;
  dns: boolean;
  portCheck: boolean;
  interfaces: boolean;
  arp: boolean;
};

export interface SystemCapabilities {
  platform: string;
  commands: SystemCapabilitiesCommands;
  demoMode: boolean;
}

export interface PingInput {
  /** @minLength 1 */
  target: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  count?: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  timeout?: number;
}

export interface TracerouteInput {
  /** @minLength 1 */
  target: string;
  /**
     * @minimum 1
     * @maximum 30
     */
  maxHops?: number;
  /**
     * @minimum 1
     * @maximum 60
     */
  timeout?: number;
}

export type DnsInputRecordType = typeof DnsInputRecordType[keyof typeof DnsInputRecordType];


export const DnsInputRecordType = {
  A: 'A',
  AAAA: 'AAAA',
  MX: 'MX',
  NS: 'NS',
  TXT: 'TXT',
  CNAME: 'CNAME',
  PTR: 'PTR',
  SOA: 'SOA',
} as const;

export interface DnsInput {
  /** @minLength 1 */
  target: string;
  recordType?: DnsInputRecordType;
  /** @nullable */
  server?: string | null;
}

export interface PortCheckInput {
  /** @minLength 1 */
  target: string;
  /**
     * @minimum 1
     * @maximum 65535
     */
  port: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  timeout?: number;
}

export type DiagnosticRunStatus = typeof DiagnosticRunStatus[keyof typeof DiagnosticRunStatus];


export const DiagnosticRunStatus = {
  success: 'success',
  failure: 'failure',
  partial: 'partial',
  running: 'running',
} as const;

export type ParsedResultDetails = { [key: string]: unknown };

export interface ParsedResult {
  summary: string;
  success: boolean;
  details?: ParsedResultDetails;
}

export interface RunMetrics {
  /** @nullable */
  latencyMin?: number | null;
  /** @nullable */
  latencyMax?: number | null;
  /** @nullable */
  latencyAvg?: number | null;
  /** @nullable */
  packetLoss?: number | null;
  /** @nullable */
  jitter?: number | null;
  /** @nullable */
  hopCount?: number | null;
  /** @nullable */
  packetsTransmitted?: number | null;
  /** @nullable */
  packetsReceived?: number | null;
}

export interface RouteHop {
  hop: number;
  /** @nullable */
  host?: string | null;
  /** @nullable */
  ip?: string | null;
  /** @nullable */
  rtt1?: number | null;
  /** @nullable */
  rtt2?: number | null;
  /** @nullable */
  rtt3?: number | null;
  timeout: boolean;
}

export interface DnsRecord {
  type: string;
  name: string;
  value: string;
  /** @nullable */
  ttl?: number | null;
}

export interface DiagnosticRun {
  id: number;
  type: string;
  target: string;
  status: DiagnosticRunStatus;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  durationMs?: number | null;
  demoMode: boolean;
  /** @nullable */
  rawOutput?: string | null;
  parsedResult?: ParsedResult;
  metrics?: RunMetrics;
  /** @nullable */
  hops?: RouteHop[] | null;
  /** @nullable */
  dnsRecords?: DnsRecord[] | null;
  /** @nullable */
  error?: string | null;
  parseWarnings: string[];
}

export type InterfaceStatusState = typeof InterfaceStatusState[keyof typeof InterfaceStatusState];


export const InterfaceStatusState = {
  up: 'up',
  down: 'down',
  unknown: 'unknown',
} as const;

export interface InterfaceStatus {
  name: string;
  /** @nullable */
  displayName?: string | null;
  state: InterfaceStatusState;
  /** @nullable */
  ipv4?: string | null;
  /** @nullable */
  ipv6?: string | null;
  /** @nullable */
  mac?: string | null;
  /** @nullable */
  gateway?: string | null;
  /** @nullable */
  netmask?: string | null;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  mtu?: number | null;
  /** @nullable */
  rx?: number | null;
  /** @nullable */
  tx?: number | null;
}

export interface HistoryResponse {
  runs: DiagnosticRun[];
  total: number;
  limit: number;
  offset: number;
}

export type SavedProfileConfig = { [key: string]: unknown };

export interface SavedProfile {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  targets: string[];
  diagnostics: string[];
  config?: SavedProfileConfig;
  /** @nullable */
  lastRunAt?: string | null;
  createdAt: string;
  isPreset: boolean;
}

export type ProfileInputConfig = { [key: string]: unknown };

export interface ProfileInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  description?: string | null;
  category?: string;
  targets: string[];
  diagnostics: string[];
  config?: ProfileInputConfig;
}

export type AlertSeverity = typeof AlertSeverity[keyof typeof AlertSeverity];


export const AlertSeverity = {
  info: 'info',
  warning: 'warning',
  error: 'error',
  critical: 'critical',
} as const;

export interface Alert {
  id: string;
  severity: AlertSeverity;
  message: string;
  /** @nullable */
  detail?: string | null;
  /** @nullable */
  target?: string | null;
  triggeredAt: string;
  resolved: boolean;
}

export interface TrendPoint {
  timestamp: string;
  /** @nullable */
  value?: number | null;
  /** @nullable */
  label?: string | null;
}

export interface OverviewMetrics {
  /** @nullable */
  avgLatency?: number | null;
  /** @nullable */
  packetLoss?: number | null;
  /** @nullable */
  avgJitter?: number | null;
  /** @nullable */
  dnsResolutionTime?: number | null;
  totalRuns: number;
  /** @nullable */
  successRate?: number | null;
  activeAlerts: number;
  /** @nullable */
  uptimePercent?: number | null;
  latencyTrend: TrendPoint[];
  packetLossTrend: TrendPoint[];
  recentRuns: DiagnosticRun[];
}

export interface ErrorResponse {
  error: string;
  /** @nullable */
  code?: string | null;
}

export type GetDiagnosticHistoryParams = {
/**
 * Filter by diagnostic type (ping, traceroute, dns, port-check, gateway)
 */
type?: string;
/**
 * Filter by target host
 */
target?: string;
/**
 * Filter by status (success, failure, partial)
 */
status?: string;
/**
 * ISO date string lower bound
 */
from?: string;
/**
 * ISO date string upper bound
 */
to?: string;
/**
 * Max results to return (default 50)
 */
limit?: number;
/**
 * Pagination offset
 */
offset?: number;
};