# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Network Diagnostics Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /diagnostics/traceroute:
    post:
      operationId: runTraceroute
      tags: [diagnostics]
      summary: Run traceroute diagnostic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TracerouteInput"
      responses:
        "200":
          description: Traceroute results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticRun"
...
  /diagnostics/port-check:
    post:
      operationId: runPortCheck
      tags: [diagnostics]
      summary: Check TCP port reachability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortCheckInput"
      responses:
        "200":
          description: Port check results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticRun"
...
  /diagnostics/history:
    get:
      operationId: getDiagnosticHistory
      tags: [history]
      summary: Get diagnostic run history
      parameters:
        - name: type
          in: query
          schema:
            type: string
          description: Filter by diagnostic type (ping, traceroute, dns, port-check, gateway)
        - name: target
          in: query
          schema:
            type: string
          description: Filter by target host
        - name: status
          in: query
          schema:
            type: string
          description: Filter by status (success, failure, partial)
        - name: from
          in: query
          schema:
            type: string
          description: ISO date string lower bound
        - name: to
          in: query
          schema:
            type: string
          description: ISO date string upper bound
        - name: limit
          in: query
          schema:
            type: integer
          description: Max results to return (default 50)
        - name: offset
          in: query
          schema:
            type: integer
          description: Pagination offset
      responses:
        "200":
          description: History list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryResponse"
  /diagnostics/run/{id}:
    get:
      operationId: getDiagnosticRun
      tags: [history]
      summary: Get a specific diagnostic run by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Diagnostic run details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticRun"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /diagnostics/run/{id}/rerun:
    post:
      operationId: rerunDiagnostic
      tags: [diagnostics]
      summary: Re-run a previous diagnostic
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: New diagnostic run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticRun"
  /profiles:
    get:
      operationId: getProfiles
      tags: [profiles]
      summary: List saved profiles
      responses:
        "200":
          description: Profile list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedProfile"
    post:
      operationId: createProfile
      tags: [profiles]
      summary: Create a saved profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "201":
          description: Created profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedProfile"
  /profiles/{id}:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get a profile by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProfile
      tags: [profiles]
      summary: Update a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedProfile"
    delete:
      operationId: deleteProfile
      tags: [profiles]
      summary: Delete a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /profiles/{id}/run:
    post:
      operationId: runProfile
      tags: [profiles]
      summary: Execute all diagnostics in a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Diagnostic runs from this profile
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiagnosticRun"
...
components:
  schemas:
...
    TracerouteInput:
      type: object
      properties:
        target:
          type: string
          minLength: 1
        maxHops:
          type: integer
          minimum: 1
          maximum: 30
        timeout:
          type: integer
          minimum: 1
          maximum: 60
      required:
        - target
...
    PortCheckInput:
      type: object
      properties:
        target:
          type: string
          minLength: 1
        port:
          type: integer
          minimum: 1
          maximum: 65535
        timeout:
          type: integer
          minimum: 1
          maximum: 30
      required:
        - target
        - port
    DiagnosticRun:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        target:
          type: string
        status:
          type: string
          enum: [success, failure, partial, running]
        startedAt:
          type: string
        completedAt:
          type: ["string", "null"]
        durationMs:
          type: ["integer", "null"]
        demoMode:
          type: boolean
        rawOutput:
          type: ["string", "null"]
        parsedResult:
          $ref: "#/components/schemas/ParsedResult"
        metrics:
          $ref: "#/components/schemas/RunMetrics"
        hops:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/RouteHop"
        dnsRecords:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/DnsRecord"
        error:
          type: ["string", "null"]
        parseWarnings:
          type: array
          items:
            type: string
      required:
        - id
        - type
        - target
        - status
        - startedAt
        - demoMode
        - parseWarnings
    ParsedResult:
      type: object
      properties:
        summary:
          type: string
        success:
          type: boolean
        details:
          type: object
      required:
        - summary
        - success
    RunMetrics:
      type: object
      properties:
        latencyMin:
          type: ["number", "null"]
        latencyMax:
          type: ["number", "null"]
        latencyAvg:
          type: ["number", "null"]
        packetLoss:
          type: ["number", "null"]
        jitter:
          type: ["number", "null"]
        hopCount:
          type: ["integer", "null"]
        packetsTransmitted:
          type: ["integer", "null"]
        packetsReceived:
          type: ["integer", "null"]
    RouteHop:
      type: object
      properties:
        hop:
          type: integer
        host:
          type: ["string", "null"]
        ip:
          type: ["string", "null"]
        rtt1:
          type: ["number", "null"]
        rtt2:
          type: ["number", "null"]
        rtt3:
          type: ["number", "null"]
        timeout:
          type: boolean
      required:
        - hop
        - timeout
    DnsRecord:
      type: object
      properties:
        type:
          type: string
        name:
          type: string
        value:
          type: string
        ttl:
          type: ["integer", "null"]
      required:
        - type
        - name
        - value
...
    HistoryResponse:
      type: object
      properties:
        runs:
          type: array
          items:
            $ref: "#/components/schemas/DiagnosticRun"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - runs
        - total
        - limit
        - offset
    SavedProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        targets:
          type: array
          items:
            type: string
        diagnostics:
          type: array
          items:
            type: string
        config:
          type: object
        lastRunAt:
          type: ["string", "null"]
        createdAt:
          type: string
        isPreset:
          type: boolean
      required:
        - id
        - name
        - category
        - targets
        - diagnostics
        - createdAt
        - isPreset
    ProfileInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        category:
          type: string
        targets:
          type: array
          items:
            type: string
        diagnostics:
          type: array
          items:
            type: string
        config:
          type: object
      required:
        - name
        - targets
        - diagnostics
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: ["string", "null"]
      required:
        - error